move authorizationList component to bh
