+++ src/views/repository/Settings.vue
@@ -59,7 +59,7 @@ import RepositoryBase from './Base';
 import EditProfileForm from '@/components-v1/user/EditProfileForm';
 import EditRepositoryForm from '@/components-v1/repository/EditRepositoryForm';
 import SetAuthorizationRoleForm from '@/components-v1/repository/SetAuthorizationRoleForm';
-import AuthorizationsList from '@/components-v1/repository/AuthorizationsList';
+import AuthorizationsList from '@/components/repository/AuthorizationsList';
 import AuthorizationRequestsList from '@/components-v1/repository/AuthorizationRequestsList';
 import LoginForm from '@/components-v1/auth/LoginForm';
